Pull out more and minor reformatting
